Update builtinfunction.Roles.md
CheckRole function returns false if user has specified role but is deactivated. It makes sense when this function is used for access check but I actually encountered this behaviour when failed to update some of already existing value for users with specific roles. Turns out for Users that are not isActive function always returns false.